CF: Update deploy scripts.

+++ scripts/aws/02_install_codebase.sh
@@ -1,8 +1,13 @@
+#!/usr/bin/env bash
 # Delete the /var/www directory, recreate it, and go to it
-sudo su
-sudo fusermount -u /var/www/s3/
-exit
-sudo rm -rf /var/www
+if [ -d "/var/www/s3/" ]; then
+    sudo su
+    sudo fusermount -u /var/www/s3/
+    exit
+fi
+if [ -d "/var/www/" ]; then
+    sudo rm -rf /var/www
+fi
 sudo mkdir -p /var/www
 cd /var/www
